Draft hoverable tools, modularize settings UI
