+++ src/js/tools/tools.js
@@ -1,5 +1,4 @@
-import {slug} from '../lib';
-import settings from '../tools/settings';
+import {getSettings} from './settings-ui';
 import version from '../version';
 
 const style = `
@@ -17,6 +16,7 @@ const style = `
 
     #tools {
       position: absolute;
+      width: 120px;
       right: 32px;
       top: 16px;
       z-index: 8000;
@@ -104,7 +104,7 @@ const style = `
       position: absolute;
       width: 150px;
       top: -8px;
-      right: 111px;
+      right: 145px;
       z-index: 8000;
       background: white;
       margin: 0;
@@ -258,12 +258,52 @@ function closeTools() {
   document.querySelector('#tools').style.display = 'none';
 }
 
+/**
+ * As needed, hide tool panels that are triggered by hovering
+ */
+function handleHideForHoverables(trigger, tool, toolHeader, toolHeaders) {
+  if (trigger === 'mouseenter') {
+
+    // Hide panel when hover leaves tool header, unless hovering to panel
+    toolHeader.addEventListener('mouseleave', event => {
+      const toElement = event.toElement;
+      const toId = toElement.id;
+      const panelElement = document.querySelector('.ideo-tool-panel');
+      if (
+        panelElement && !panelElement.contains(toElement) &&
+        toId !== tool
+      ) {
+        console.log('deactivate from handleHideForHoverables')
+        deactivate(toolHeaders);
+        panelElement.remove();
+      }
+
+      console.log('')
+      console.log('')
+      console.log('')
+    });
+
+    // if ()
+  }
+}
+
+/** Determine action that should trigger a tool panel to display */
+function getTrigger(toolHeader) {
+  const shouldHover =
+    Array.from(toolHeader.classList).includes('ideo-tool-hover');
+  const trigger = shouldHover ? 'mouseenter' : 'click';
+  return trigger;
+}
+
 /** Shows clicked tool as active, displays resulting panel */
 function handleToolClick(ideo) {
   const toolHeaders = document.querySelectorAll('#tools > ul > li');
 
   toolHeaders.forEach(toolHeader => {
-    toolHeader.addEventListener('click', event => {
+
+    const trigger = getTrigger(toolHeader);
+
+    toolHeader.addEventListener(trigger, event => {
 
       // Show only clicked tool header as active
       deactivate(toolHeaders);
@@ -273,6 +313,8 @@ function handleToolClick(ideo) {
       const panel = getPanel(tool, ideo);
 
       document.querySelector('#gear').insertAdjacentHTML('beforeend', panel);
+
+      handleHideForHoverables(trigger, tool, toolHeader, toolHeaders);
     });
   });
 
@@ -298,150 +340,6 @@ function handleGearClick(ideo) {
   handleToolClick(ideo);
 
   handleSettingsHeaderClick(ideo);
-}
-
-/** Ensure string value can be rendered as an HTML "title" attribute */
-function toTitle(value) {
-  if (typeof value !== 'undefined') {
-    return value
-      .trim()
-      .replace(/"/g, '&quot;')
-      .replace(/\n/g, '')
-      .replace(/\s{2,}/g, ' ');
-  }
-}
-
-function getIdAttr(setting) {
-  const id = 'id' in setting ? setting.id : slug(setting.name);
-  return `setting-${slug(id)}`;
-}
-
-/** Get HTML label for setting header */
-function getHeader(setting, option=null, optionId=null) {
-  // Get a header for each setting
-
-  let name; let id; let title;
-
-  if (option) {
-    name = option;
-    id = optionId;
-    title = '';
-  } else {
-    name = 'shortName' in setting ? setting.shortName : setting.name;
-    id = getIdAttr(setting);
-    title = ` title="${toTitle(setting.description)}"`;
-  }
-
-  const underline = option === null ? '' : ' no-underline';
-
-  const attrs =
-    `class="setting${underline}" for="${id}"${title}`;
-
-  return `<label ${attrs}>${name}</label>`;
-}
-
-/** Transform options to an array of list items (<li>'s) */
-function getOptions(setting, name) {
-
-  const typeAttr = `type="${setting.type}"`;
-
-  if ('options' in setting === false) {
-    // type="number"
-    const id = getIdAttr(setting);
-    return `<input ${typeAttr} id="${id}"/><br/>`;
-  }
-
-  if (Array.isArray(setting.options) === false) return '';
-
-  const description = toTitle(setting.description);
-
-  if (setting.type === 'radio') {
-    return setting.options.map(option => {
-
-      const id = 'setting-' + slug(option);
-      const attrs = `${typeAttr} id="${id}" title="${description}"`;
-
-      const input = `<input ${attrs} name="${name}" value="${id}"/>`;
-      const label = getHeader(setting, option, id);
-      const item = '<br/>' + input + label;
-      return item;
-    }).join('') + '<br/>';
-  }
-}
-
-function listTabs(themeObj, i) {
-  const settingsList = themeObj.list
-    .map(area => {
-
-      const areaHeading = `<div class="area-header">${area.area}</div>`;
-
-      const settingsByArea = area.settings
-        .filter(setting => {
-          const displayedTypes = ['string', 'number', 'checkbox', 'radio'];
-          return displayedTypes.includes(setting.type);
-        })
-        .map(setting => {
-          const name =
-            ('id' in setting) ? setting.id : slug(setting.name);
-
-          const header = getHeader(setting);
-          const options = getOptions(setting, name);
-
-          return '<div>' + header + options + '</div>';
-        }).join('');
-
-      return (
-        areaHeading +
-        '<div class="area-content">' + settingsByArea + '</div>'
-      );
-    }).join('<br/>');
-
-  const theme = themeObj.theme;
-  const activeClass = (i === 0) ? ' class="active"' : '';
-
-  const htmlId = `${slug(theme)}-tab`;
-
-  return `
-    <div id="${htmlId}"${activeClass}>
-      ${settingsList}
-    </div>`;
-}
-
-/**
- * Get list of configurable Ideogram settings; each has a header and options
- *
- * Settings are grouped by theme and area.
- *
- * -> Theme (e.g. Basic, Chromosomes )
- * |-> Area (e.g. Biology, Data)
- *  |-> Setting (e.g. Organism, Height)
- *
- * @param {Array} settingThemes
- */
-function list(settingThemes) {
-
-  const navHeaders = settingThemes.map((themeObj, i) => {
-    const activeClass = (i === 0) ? ' active"' : '';
-    const theme = themeObj.theme;
-    return `
-      <li class="ideo-settings-header ${activeClass}">
-        <a href="#${slug(theme)}-tab">${theme}</a>
-      </li>`;
-  }).join('');
-
-  const nav = `
-    <div class="tab-panel">
-    <ul class="nav">
-      ${navHeaders}
-    </ul>`;
-
-  const tabs = settingThemes.map((themeObj, i) => {
-    return listTabs(themeObj, i);
-  }).join('');
-
-  const tabContent = `<div class="tab-content">${tabs}</div>`;
-
-  return nav + tabContent;
 }
 
 function showGearOnIdeogramHover(ideo) {
@@ -493,14 +391,6 @@ function getAbout() {
     </div>`;
 }
 
-function getSettings() {
-  const settingsList = list(settings);
-  return `
-    <ul id="settings">
-        ${settingsList}
-    </ul>`;
-}
-
 function hideOnClickOutside(selector) {
   const elements = document.querySelectorAll('#gear, #tools');
   const outsideClickListener = event => {
@@ -525,12 +415,14 @@ function hideOnClickOutside(selector) {
 
 function initTools(ideo) {
 
+  const triangle = '<span style="float: right">&blacktriangleright;</span>';
+
   const toolsHtml = `
     ${style}
     <div id="gear" style="display: none">${gearIcon}</div>
     <div id="tools" style="display: none">
       <ul>
-        <li id="download-tool">Download</li>
+        <li id="download-tool" class="ideo-tool-hover">Download ${triangle}</li>
         <li id="about-tool">About</li>
       </ul>
     </div>`;
